backend/requirements.txt
Django>=4.2,<5.0
djangorestframework
psycopg2-binary
celery
redis
django-cors-headers
django-celery-results
gunicorn
python-dotenv
djangorestframework-simplejwt
